% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_fqa.R
\name{assess_fqa}
\alias{assess_fqa}
\title{Obtain Study Information From Universal FQA Inventory as a Data Frame}
\usage{
assess_fqa(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_fqa() or other similar function}
}
\value{
A data frame with 9 columns:
\itemize{
   \item Scientific Name (character)
   \item Family (character)
   \item Acronym (character)
   \item Native? (character)
   \item C (numeric)
   \item W (numeric)
   \item Physiognomy (character)
   \item Duration (character)
   \item Common Name (character)
}
}
\description{
Obtain Study Information From Universal FQA Inventory as a Data Frame
}
\examples{
\dontrun{
## assess_fqa can be used with a .csv file downloaded from the universal FQA website:

assess_fqa(open_dunes)

## or with a download function:

assess_fqa(download_fqa(25002))

## assess_fqa can also be used with saved data from a download function:

df <- download_fqa(25002)
assess_fqa(df)
}
}
