% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_fqa.R
\name{glance_fqa}
\alias{glance_fqa}
\title{Obtain Overview Information From a Universal FQA Inventory as a Data Frame}
\usage{
glance_fqa(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_fqa() or other similar function}
}
\value{
A data frame with 52 columns:
\itemize{
   \item Title (character)
   \item Date (POSIXct)
   \item Site Name (character)
   \item City (character)
   \item County (character)
   \item State (character)
   \item Country (character)
   \item FQA DB Region (character)
   \item FQA DB Publication Year (character)
   \item FQA DB Description (character)
   \item Practitioner (character)
   \item Latitude (character)
   \item Longitude (character)
   \item Weather Notes (character)
   \item Duration Notes (character)
   \item Community Type Notes (character)
   \item Other Notes (character)
   \item Private/Public (character)
   \item Total Mean C (numeric)
   \item Native Mean C (numeric)
   \item Total FQI (numeric)
   \item Native FQI (numeric)
   \item Adjusted FQI (numeric)
   \item % C value 0 (numeric)
   \item % C value 1-3 (numeric)
   \item % C value 4-6 (numeric)
   \item % C value 7-10 (numeric)
   \item Native Tree Mean C (numeric)
   \item Native Shrub Mean C (numeric)
   \item Native Herbaceous Mean C (numeric)
   \item Species Richness (numeric)
   \item Total Species (numeric)
   \item Native Species (numeric)
   \item Non-native Species (numeric)
   \item Species Wetness (numeric)
   \item Mean Wetness (numeric)
   \item Native Mean Wetness (numeric)
   \item Tree (numeric)
   \item Shrub (numeric)
   \item Vine (numeric)
   \item Forb (numeric)
   \item Grass (numeric)
   \item Sedge (numeric)
   \item Rush (numeric)
   \item Fern (numeric)
   \item Bryophyte (numeric)
   \item Annual (numeric)
   \item Perennial (numeric)
   \item Biennial (numeric)
   \item Native Annual (numeric)
   \item Native Perennial: (numeric)
   \item Native Biennial: (numeric)
}
}
\description{
Obtain Overview Information From a Universal FQA Inventory as a Data Frame
}
\examples{
\dontrun{
## glance_fqa can be used with a .csv file downloaded from the universal FQA website:

glance_fqa(open_dunes)

## or with a download function:

glance_fqa(download_fqa(25002))

## glance_fqa can also be used with saved data from a download function:

df <- download_fqa(25002)
glance_fqa(df)
}

}
