% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_phis_transect.R
\name{glance_phis}
\alias{glance_phis}
\title{Obtain Physiognomy Information From a Universal FQA Transect as a Data Frame}
\usage{
glance_phis(data_set)
}
\arguments{
\item{data_set}{a data frame downloaded from Universal FQA using download_transect() or other similar function}
}
\value{
A data frame with 6 columns:
\itemize{
   \item Physiognomy (character)
   \item Frequency (numeric)
   \item Coverage (numeric)
   \item Relative Frequency % (numeric)
   \item Relative Coverage % (numeric)
   \item Relative Importance Value (numeric)
}
}
\description{
Obtain Physiognomy Information From a Universal FQA Transect as a Data Frame
}
\examples{
\dontrun{
## glance_phis can be used with a download function:

glance_phis(download_transect(6325))

## glance_phis can also be used with saved data from a download function:

df <- download_transect(6325)
glance_phis(df)
}
}
